move getComment to JsonLDUtil
